chore: add usize to simplify casts later on

+++ crates/comm/src/packet.rs
@@ -132,7 +132,7 @@ impl NetworkPacket {
         // read packet header
         let packet_type = reader.read_u8()?.into();
         let seq_number = reader.read_u32::<BigEndian>()?;
-        let compressed_data_length = reader.read_u32::<BigEndian>()?;
+        let compressed_data_length = reader.read_u32::<BigEndian>()? as usize;
         let uncompressed_data_length = reader.read_u32::<BigEndian>()?;
 
         if (uncompressed_data_length as usize) == 0 {
